+++ src/utils/sets/difference.ts
@@ -1,5 +1,7 @@
 /**
  * Create difference set of two sets.
+ *
+ * @deprecated use new javascript set methods instead @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set
  */
 export function difference<TItem>(
     a: Set<TItem>,